Move password recovery to new views folder
