--- apps/web/src/components/recovery/form/styles.module.css
@@ -1,13 +1,0 @@
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-  text-align: center;
-}
-
-.group {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 1rem;
-}
