--- apps/web/src/components/recovery/header/styles.module.css
@@ -1,12 +1,0 @@
-.header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.5rem;
-  text-align: center;
-}
-
-.logo {
-  width: 4rem;
-  height: 4rem;
-}
