--- apps/web/src/components/recovery/styles.module.css
@@ -1,8 +1,0 @@
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-  width: 100%;
-  max-width: 40rem;
-  padding: 3rem;
-}
